refactor(monitor): Extract conflictSet into its own file
Move the generic conflictSet, conflictSetFacts, and conflictSetTerms
helpers out of monitor.go into a new monitor/conflict_set.go. The
Unifier interface they parameterise over stays in monitor.go because
it is also used by handleHints/handleTriggers.

Pure refactor; no behavioural change. This isolates the conflict-set
matching logic so the upcoming indexing rewrite has a self-contained
file to land in.
